+++ sdk/communication/communication-messages-rest/samples/v2/typescript/src/GetMessageTemplateList.ts
@@ -14,29 +14,29 @@ import * as dotenv from "dotenv";
 dotenv.config();
 
 async function main(): Promise<void> {
-    const credential = new AzureKeyCredential(process.env.ACS_ACCESS_KEY || "");
-    const endpoint = process.env.ACS_URL || "";
-    const client = MessageTemplateClient(endpoint, credential);
-    console.log("Fetch Templates...");
-    const response = await client.path("/messages/channels/{channelId}/templates", process.env.CHANNEl_ID || "")
+  const credential = new AzureKeyCredential(process.env.ACS_ACCESS_KEY || "");
+  const endpoint = process.env.ACS_URL || "";
+  const client = MessageTemplateClient(endpoint, credential);
+  console.log("Fetch Templates...");
+  const response = await client.path("/messages/channels/{channelId}/templates", process.env.CHANNEL_ID || "")
     .get({
-        queryParameters: { maxPageSize: 2 }
+      queryParameters: { maxPageSize: 2 }
     });
 
-    if(isUnexpected(response)) {
-        throw new Error("Failed to get template for the channel.");
-    }
+  if (isUnexpected(response)) {
+    throw new Error("Failed to get template for the channel.");
+  }
 
-    // The paginate helper creates a paged async iterator using metadata from the first page.
-    const items = paginate(client, response);
+  // The paginate helper creates a paged async iterator using metadata from the first page.
+  const items = paginate(client, response);
 
-    // We get an PageableAsyncIterator so we need to do `for await`.
-    for await (const item of items) {
-        console.log(JSON.stringify(item, null, 2));
-    }    
+  // We get an PageableAsyncIterator so we need to do `for await`.
+  for await (const item of items) {
+    console.log(JSON.stringify(item, null, 2));
+  }
 }
 
 main().catch((error) => {
-    console.error("Encountered an error while sending message: ", error);
-    throw error;
+  console.error("Encountered an error while sending message: ", error);
+  throw error;
 });