+++ sdk/cosmosdb/cosmos/test/public/functional/NonStreamingQueryPolicy.spec.ts
@@ -559,8 +559,7 @@ async function executeQueryAndVerifyOrder(
   assert.equal(count, size);
 }
 
-// Skipping these tests as they are not supported by public emulator
-describe.skip("Full text search feature", async () => {
+describe("Full text search feature", async () => {
   let database: Database;
 
   beforeAll(async () => {